requirements.txt
astroid==2.3.3
attrs==19.3.0
bitmex==0.2.2
boto3==1.10.26
botocore==1.13.28
bravado==10.4.3
bravado-core==5.16.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
docutils==0.15.2
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.0.0
msgpack-python==0.5.6
poloniexapi==0.5.7
pylint==2.4.4
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rfc3987==1.3.8
s3transfer==0.2.1
simplejson==3.17.0
six==1.13.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
webcolors==1.10
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
gunicorn==19.3.0